Shitty patch for #67.
Note block markdown not respecting occlusion by another block. This is caused because the Window Child doesn't use the draw list of the flowgraph and thus does not respect the order of elements.

To make this work, I created a new child flag to force the child to use the parent window draw list.

This method has its problems. Clipping is broken while scrolling but not very noticeable. Whole be nice to find a better way.